reimplement PublisherBanner for iOS
please note, these changes are not backwards compatible.

The exported view properties now match the properties of the DFPBannerView.

I've now implemented the `validAdSizes` property, and replaced the `testDeviceID` property with `testDevices`. Also the simulator device id can be specified using `PublisherBanner.simulatorId` instead of `"EMULATOR"`.

The conversion of json to GADAdSize is now implemented using a RCTConvert category.

And lastly, the `loadBanner` method is now exposed to react, to allow easy banner refresh, and removes the need of checking which properties are available before making a request to admob.